feat: Capability-based security evolution (v0.9.0) — audit + implementation
Architecture audit of proposed Capability Token + Semantic Gateway model:
- Rejected HMAC-SHA256 in capabilities (kernel object references are sufficient)
- Evolved existing 64-bit bitmask ACL to dual-mode (process-level + object-level)
- Added SubsystemId.AI for AI capsule support
- MAX_TASKS increased from 8 to 64

New modules:
- capability.zig: Derivation (subset-only), checking with TTL, revocation, snapshot store
- policy_engine.zig: Kernel-level policy evaluation (whitelist model, 8 default rules)
- ipc.zig: Zircon-inspired Channel-based IPC (bounded message queues, cap-passing)
- ai_capsule.zig: AI capsule lifecycle (create/start/update/rollback/stop)

Modified modules:
- kernel_integrate.zig: New cap bits (AI_RUNTIME/MANAGE/POLICY_SET/AUDIT_READ),
  PCB fields (cap_expire_tick, memory_quota, cpu_quota), TTL expiry in polerAuthenticate(),
  public capability mapping functions
- object_manager.zig: access_mask enforcement, cap_generation/cap_revoked fields,
  checkHandleAccess(), revokeHandle(), global singleton wrappers
- scheduler.zig: MAX_TASKS=64, reduced debug verbosity
- subsystem.zig: SubsystemId.AI, AI_SYSCALL_BASE (0x3000+), requiredCapsForSyscall()
- main64.zig: Integration of all new modules in boot sequence

Build: 0 errors, ~2MB binary

+++ zig-kernel/src64/subsystem/posix/posix_api.zig
@@ -717,7 +717,7 @@ fn posixOpen(path_addr: u64, flags: i32, _mode: i32) i64 {
 
         // Also create an Object Manager handle
         if (objmgr_ref) |om| {
-            const access: u64 = if (!writable) 0x80000000 else if (!readable) 0x40000000 else 0xC0000000;
+            const access: u32 = if (!writable) 0x80000000 else if (!readable) 0x40000000 else 0xC0000000;
             const handle = om.createHandle(.File, access);
             if (fd_table.getFd(fd)) |entry| {
                 entry.obj_handle = handle;
@@ -734,7 +734,7 @@ fn posixOpen(path_addr: u64, flags: i32, _mode: i32) i64 {
 
     const fd = fd_table.allocFd() orelse return -@as(i64, subsys.EMFILE);
     if (objmgr_ref) |om| {
-        const access: u64 = if (!writable) 0x80000000 else if (!readable) 0x40000000 else 0xC0000000;
+        const access: u32 = if (!writable) 0x80000000 else if (!readable) 0x40000000 else 0xC0000000;
         const handle = om.createHandle(.File, access);
         if (fd_table.getFd(fd)) |entry| {
             entry.obj_handle = handle;
